JAVA-2697: Extract CommandEventSender interface and move implementation to a top-level non-public class
