+++ driver-core/src/main/com/mongodb/connection/InternalStreamConnection.java
@@ -17,7 +17,6 @@
 package com.mongodb.connection;
 
 import com.mongodb.MongoClientException;
-import com.mongodb.MongoCommandException;
 import com.mongodb.MongoCompressor;
 import com.mongodb.MongoException;
 import com.mongodb.MongoInternalException;
@@ -34,13 +33,9 @@ import com.mongodb.diagnostics.logging.Loggers;
 import com.mongodb.event.CommandListener;
 import com.mongodb.session.SessionContext;
 import org.bson.BsonBinaryReader;
-import org.bson.BsonDocument;
-import org.bson.BsonInt32;
-import org.bson.BsonValue;
 import org.bson.ByteBuf;
 import org.bson.codecs.BsonDocumentCodec;
 import org.bson.codecs.Decoder;
-import org.bson.codecs.RawBsonDocumentCodec;
 import org.bson.io.ByteBufferBsonInput;
 
 import java.io.IOException;
@@ -63,9 +58,6 @@ import static com.mongodb.connection.ProtocolHelper.getCommandFailureException;
 import static com.mongodb.connection.ProtocolHelper.getMessageSettings;
 import static com.mongodb.connection.ProtocolHelper.getOperationTime;
 import static com.mongodb.connection.ProtocolHelper.isCommandOk;
-import static com.mongodb.connection.ProtocolHelper.sendCommandFailedEvent;
-import static com.mongodb.connection.ProtocolHelper.sendCommandStartedEvent;
-import static com.mongodb.connection.ProtocolHelper.sendCommandSucceededEvent;
 import static com.mongodb.internal.async.ErrorHandlingResultCallback.errorHandlingCallback;
 import static java.lang.String.format;
 import static java.util.Arrays.asList;
@@ -239,7 +231,7 @@ class InternalStreamConnection implements InternalConnection {
     public <T> T sendAndReceive(final CommandMessage message, final Decoder<T> decoder, final SessionContext sessionContext) {
         ByteBufferBsonOutput bsonOutput = new ByteBufferBsonOutput(this);
         LazyCommandDocument lazyCommandDocument = new LazyCommandDocument(message, bsonOutput);
-        CommandEventSender commandEventSender = new CommandEventSender(message, lazyCommandDocument);
+        CommandEventSender commandEventSender = createCommandEventSender(message, lazyCommandDocument);
 
         try {
             message.encode(bsonOutput, sessionContext);
@@ -262,7 +254,6 @@ class InternalStreamConnection implements InternalConnection {
             throw e;
         }
     }
-
     private void sendCommandMessage(final CommandMessage message, final LazyCommandDocument lazyCommandDocument,
                                     final ByteBufferBsonOutput bsonOutput, final SessionContext sessionContext) {
         try {
@@ -292,7 +283,7 @@ class InternalStreamConnection implements InternalConnection {
             updateSessionContext(sessionContext, responseBuffers);
 
             if (!isCommandOk(responseBuffers)) {
-                throw getCommandFailureException(getResponseDocument(responseBuffers, message.getId(), new BsonDocumentCodec()),
+                throw getCommandFailureException(responseBuffers.getResponseDocument(message.getId(), new BsonDocumentCodec()),
                         description.getServerAddress());
             }
 
@@ -320,7 +311,7 @@ class InternalStreamConnection implements InternalConnection {
         try {
             message.encode(bsonOutput, sessionContext);
             LazyCommandDocument lazyCommandDocument = new LazyCommandDocument(message, bsonOutput);
-            CommandEventSender commandEventSender = new CommandEventSender(message, lazyCommandDocument);
+            CommandEventSender commandEventSender = createCommandEventSender(message, lazyCommandDocument);
             commandEventSender.sendStartedEvent();
 
             if (sendCompressor == null || SECURITY_SENSITIVE_COMMANDS.contains(lazyCommandDocument.getName())) {
@@ -370,8 +361,9 @@ class InternalStreamConnection implements InternalConnection {
                                         isCommandOk(new BsonBinaryReader(new ByteBufferBsonInput(responseBuffers.getBodyByteBuffer())));
                                 responseBuffers.reset();
                                 if (!commandOk) {
-                                    MongoException commandFailureException = getCommandFailureException(getResponseDocument(responseBuffers,
-                                            messageId, new BsonDocumentCodec()), description.getServerAddress());
+                                    MongoException commandFailureException = getCommandFailureException(
+                                            responseBuffers.getResponseDocument(messageId, new BsonDocumentCodec()),
+                                            description.getServerAddress());
                                     commandEventSender.sendFailedEvent(commandFailureException);
                                     throw commandFailureException;
                                 }
@@ -583,13 +575,6 @@ class InternalStreamConnection implements InternalConnection {
         return stream.getBuffer(size);
     }
 
-    private static <T extends BsonDocument> T getResponseDocument(final ResponseBuffers responseBuffers,
-                                                                  final int messageId, final Decoder<T> decoder) {
-        ReplyMessage<T> replyMessage = new ReplyMessage<T>(responseBuffers, decoder, messageId);
-        responseBuffers.reset();
-        return replyMessage.getDocuments().get(0);
-    }
-
     private class MessageHeaderCallback implements SingleResultCallback<ByteBuf> {
         private final SingleResultCallback<ResponseBuffers> callback;
 
@@ -656,156 +641,12 @@ class InternalStreamConnection implements InternalConnection {
         }
     }
 
-    private static final Logger COMMAND_LOGGER = Loggers.getLogger("protocol.command");
-
-    private class CommandEventSender {
-        private final long startTimeNanos;
-        private final CommandMessage message;
-        private final LazyCommandDocument lazyCommandDocument;
-
-        CommandEventSender(final CommandMessage message, final LazyCommandDocument lazyCommandDocument) {
-            this.startTimeNanos = System.nanoTime();
-            this.message = message;
-            this.lazyCommandDocument = lazyCommandDocument;
-        }
-
-        public void sendStartedEvent() {
-            if (sendRequired()) {
-                BsonDocument commandDocumentForEvent = (SECURITY_SENSITIVE_COMMANDS.contains(lazyCommandDocument.getName()))
-                        ? new BsonDocument() : lazyCommandDocument.getDocument();
-
-                if (loggingRequired()) {
-                    COMMAND_LOGGER.debug(
-                            format("Sending command {%s : %s, ...} with request id %d to database %s on connection [%s] to server %s",
-                                    lazyCommandDocument.getName(), lazyCommandDocument.getFirstValue(), message.getId(),
-                                    message.getNamespace().getDatabaseName(), description.getConnectionId(),
-                                    description.getServerAddress()));
-                }
-
-                if (eventRequired()) {
-                    sendCommandStartedEvent(message, message.getNamespace().getDatabaseName(),
-                            lazyCommandDocument.getName(), commandDocumentForEvent, getDescription(), commandListener);
-                }
-            }
-        }
-
-        public void sendFailedEvent(final Throwable t) {
-            if (sendRequired()) {
-                Throwable commandEventException = t;
-                if (t instanceof MongoCommandException && (SECURITY_SENSITIVE_COMMANDS.contains(lazyCommandDocument.getName()))) {
-                    commandEventException = new MongoCommandException(new BsonDocument(), description.getServerAddress());
-                }
-                long elapsedTimeNanos = System.nanoTime() - startTimeNanos;
-
-                if (loggingRequired()) {
-                    COMMAND_LOGGER.debug(
-                            format("Execution of command with request id %d failed to complete successfully in %.2f ms on connection [%s] "
-                                            + "to server %s",
-                                    message.getId(), nanosToMillis(elapsedTimeNanos), description.getConnectionId(),
-                                    description.getServerAddress()),
-                            commandEventException);
-                }
-
-                if (eventRequired()) {
-                    sendCommandFailedEvent(message, lazyCommandDocument.getName(), description, elapsedTimeNanos, commandEventException,
-                            commandListener);
-                }
-            }
-        }
-
-        public void sendSucceededEvent(final ResponseBuffers responseBuffers) {
-            if (sendRequired()) {
-                BsonDocument responseDocumentForEvent = (SECURITY_SENSITIVE_COMMANDS.contains(lazyCommandDocument.getName()))
-                        ? new BsonDocument()
-                        : getResponseDocument(responseBuffers, message.getId(),
-                        new RawBsonDocumentCodec());
-                long elapsedTimeNanos = System.nanoTime() - startTimeNanos;
-
-                if (loggingRequired()) {
-                    COMMAND_LOGGER.debug(
-                            format("Execution of command with request id %d completed successfully in %.2f ms on connection [%s] "
-                                            + "to server %s",
-                                    message.getId(), nanosToMillis(elapsedTimeNanos), description.getConnectionId(),
-                                    description.getServerAddress()));
-                }
-
-                if (eventRequired()) {
-                    sendCommandSucceededEvent(message, lazyCommandDocument.getName(), responseDocumentForEvent, description,
-                            elapsedTimeNanos, commandListener);
-                }
-            }
-        }
-
-        public void sendSucceededEventForOneWayCommand() {
-            if (sendRequired()) {
-                BsonDocument responseDocumentForEvent = new BsonDocument("ok", new BsonInt32(1));
-                long elapsedTimeNanos = System.nanoTime() - startTimeNanos;
-
-                if (loggingRequired()) {
-                    COMMAND_LOGGER.debug(
-                            format("Execution of one-way command with request id %d completed successfully in %.2f ms on connection [%s] "
-                                            + "to server %s",
-                                    message.getId(), nanosToMillis(elapsedTimeNanos), description.getConnectionId(),
-                                    description.getServerAddress()));
-                }
-
-                if (eventRequired()) {
-                    sendCommandSucceededEvent(message, lazyCommandDocument.getName(), responseDocumentForEvent, description,
-                            elapsedTimeNanos, commandListener);
-                }
-            }
-        }
-
-        private boolean sendRequired() {
-            return (eventRequired() || loggingRequired()) && opened();
-        }
-
-        private boolean loggingRequired() {
-            return COMMAND_LOGGER.isDebugEnabled();
-        }
-
-        private boolean eventRequired() {
-            return commandListener != null;
-        }
-
-        private double nanosToMillis(final long elapsedTimeNanos) {
-            return elapsedTimeNanos / 1000000.0;
-        }
-    }
-
-    // Lazily determine the command document and command name, since they're only needed if either a command listener or compression
-    // is enabled
-    private static final class LazyCommandDocument {
-        private final CommandMessage commandMessage;
-        private final ByteBufferBsonOutput bsonOutput;
-        private BsonDocument commandDocument;
-        private String commandName;
-        private BsonValue firstValue;
-
-        private LazyCommandDocument(final CommandMessage commandMessage, final ByteBufferBsonOutput bsonOutput) {
-            this.commandMessage = commandMessage;
-            this.bsonOutput = bsonOutput;
-        }
-
-        public String getName() {
-            if (commandName == null) {
-                commandName = getDocument().getFirstKey();
-            }
-            return commandName;
-        }
-
-        public BsonValue getFirstValue() {
-            if (firstValue == null) {
-                firstValue = getDocument().getFirstValue();
-            }
-            return firstValue;
-        }
-
-        private BsonDocument getDocument() {
-            if (commandDocument == null) {
-                commandDocument = commandMessage.getCommandDocument(bsonOutput);
-            }
-            return commandDocument;
+    private CommandEventSender createCommandEventSender(final CommandMessage message,
+                                                               final LazyCommandDocument lazyCommandDocument) {
+        if (opened() && LoggingCommandEventSender.isRequired(commandListener)) {
+            return new LoggingCommandEventSender(SECURITY_SENSITIVE_COMMANDS, description, commandListener, message, lazyCommandDocument);
+        } else {
+            return new NoOpCommandEventSender();
         }
     }
 }
